refactor av1 packet trailer support
