+++ webrtc-sys/include/livekit/packet_trailer.h
@@ -78,6 +78,14 @@ struct PacketTrailerMetadata {
   uint32_t ssrc;  // SSRC that produced this entry (for simulcast tracking)
 };
 
+/// Parses a trailer payload (TLV region followed by the trailer envelope) and
+/// returns the embedded metadata, or `std::nullopt` if the payload is missing
+/// the magic envelope or contains no recognized TLV elements.
+///
+/// Shared by the codec-agnostic trailer path and the AV1 OBU path.
+std::optional<PacketTrailerMetadata> ParseTrailerPayload(
+    webrtc::ArrayView<const uint8_t> trailer);
+
 /// Frame transformer that appends/extracts packet trailers.
 /// This transformer can be used standalone or in conjunction with e2ee.
 ///
